requirements.txt
selenium
requests
webdriver-manager
pyyaml
pytz
python-dotenv
aiofiles
aiohttp